refactor(matrix): split systolic array modules
